Reorg some logic by creating a new renderer module for each plugin. markupControllers are now responsible mainly for generating the html themselves and calling the renderers (which call up subsequent classes as needed). Also had Claude go through and re-org the import statements for clarity
